+++ src/Command/RbacDbInit.php
@@ -9,8 +9,6 @@ use Symfony\Component\Console\Command\Command;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Output\OutputInterface;
 use Yiisoft\Db\Connection\ConnectionInterface;
-use Yiisoft\Db\Schema\SchemaInterface;
-use Yiisoft\Db\Sqlite\Column;
 
 /**
  * Command for creating RBAC related database tables using Yii Database.
@@ -111,7 +109,7 @@ final class RbacDbInit extends Command
             ->createTable(
                 $this->itemsTable,
                 [
-                    'name' => (new Column(SchemaInterface::TYPE_STRING, 128))->notNull()->append('PRIMARY KEY'),
+                    'name' => 'string(128) NOT NULL PRIMARY KEY',
                     'type' => 'string(10) NOT NULL',
                     'description' => 'string(191)',
                     'ruleName' => 'string(64)',