+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # Yii RBAC Database Storage Change Log
 
-## 1.0.1 under development
+## 2.0.0 under development
 
 - New #23: Remove CLI dependencies, add `DbSchemaManager`, dump SQL for working with schema (@arogachev) 
 - Bug #24: Remove usage of `SQLite` column in DB agnostic code (@arogachev)